+++ lib/cloud_frog.dart
@@ -1,6 +1,9 @@
 /// A library to help build secure Google Cloud services using Dart Frog.
 library;
 
+export 'package:gcputil/app_check.dart';
+
+export 'src/app_check.dart';
 export 'src/auth.dart';
 export 'src/datetime.dart';
 export 'src/gcloud.dart';